+++ src/main/java/de/btu/dataglove/shared/EulerFrame.java
@@ -10,7 +10,7 @@ public class EulerFrame {
 
     public final String nameOfTask;
     public final String userName;
-    public final long timeStamp;
+    public final double timeStamp;
     public final int frameNumber;
     public final int recordingNumber;
     public final int typeOfRecording;
